% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statjobs.R
\name{pega_posts}
\alias{pega_posts}
\title{Pega posts do statjobs na web usando pacote Rfacebook}
\usage{
pega_posts(at)
}
\description{
Pegar um access_token em https://developers.facebook.com/tools/explorer
}

